Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.0.76
pydantic==2.4.2
python-dotenv==1.0.0
ultralytics>=8.0.0
